+++ sqlstore/src/main/resources/db/migration/V8501__transfer_volume.sql
@@ -1,0 +1,19 @@
+ALTER TABLE Transfer_Sample
+  ADD COLUMN distributedVolume DECIMAL(16,10),
+  ADD COLUMN distributedBoxAlias varchar(255),
+  ADD COLUMN distributedBoxPosition varchar(3);
+
+ALTER TABLE Transfer_Library
+  ADD COLUMN distributedVolume DECIMAL(16,10),
+  ADD COLUMN distributedBoxAlias varchar(255),
+  ADD COLUMN distributedBoxPosition varchar(3);
+
+ALTER TABLE Transfer_LibraryAliquot
+  ADD COLUMN distributedVolume DECIMAL(16,10),
+  ADD COLUMN distributedBoxAlias varchar(255),
+  ADD COLUMN distributedBoxPosition varchar(3);
+
+ALTER TABLE Transfer_Pool
+  ADD COLUMN distributedVolume DECIMAL(16,10),
+  ADD COLUMN distributedBoxAlias varchar(255),
+  ADD COLUMN distributedBoxPosition varchar(3);